onboarding commands
