+++ cypress/support/commands.ts
@@ -1,4 +1,5 @@
 import '@testing-library/cypress/add-commands'
+import './onboardingCommands'
 // ***********************************************
 // This example commands.js shows you how to
 // create various custom commands and overwrite
